Experiment: changing left hand letters
I've had the same layout for an entire day so obviously it's time for a
change.

I'm finding 'h' and 'o' a bit awkward, and I think I've wasted a prime
key for 'f'. This experiment keeps 'oie' together and moves 'h' onto the
home row, at the expense of 'a'.